feat: add factories to interface classes
Add factories for sync and async implementations to:
-  `Database`
-  `Replicator`
-  `Query`
-  `QueryBuilder`

Also put implementations into files that are separat from API definitions.
