+++ packages/cbl/lib/src/query/group_by.dart
@@ -1,4 +1,6 @@
+import 'proxy_query.dart';
 import 'expressions/expression.dart';
+import 'ffi_query.dart';
 import 'having.dart';
 import 'limit.dart';
 import 'order_by.dart';
