+++ packages/cbl/lib/src/query/having.dart
@@ -1,4 +1,6 @@
+import 'proxy_query.dart';
 import 'expressions/expression.dart';
+import 'ffi_query.dart';
 import 'limit.dart';
 import 'order_by.dart';
 import 'ordering.dart';
